fix(runtime): upgrade hook counters to size_t to avoid integer overflow

+++ src/core/sdk_runtime.cpp
@@ -1,7 +1,7 @@
 #include "sdk_runtime.h"
 
 #include "core/game/game_base.h"
-#include "logging/logger.hpp"
+#include "logging/logger.h" 
 
 using namespace GTASA::SDK;
 
@@ -44,8 +44,9 @@ void SDKRuntime::init()
     LOG_INFO("[SDKRuntime] Installing hooks...");
     HookRegistry::instance().sort();
 
-    uint8_t enabledHooksCount = 0;
-    uint8_t installedHooksCount = 0;
+    # FIX: Changed types from uint8_t to size_t to prevent overflow when hook count exceeds 255
+    size_t enabledHooksCount = 0;
+    size_t installedHooksCount = 0;
     for (auto& hook : HookRegistry::instance().getAll())
     {
         if (!hook->isEnabled()) continue;
@@ -53,7 +54,8 @@ void SDKRuntime::init()
         enabledHooksCount++;
     }
 
-    LOG_INFO("[SDKRuntime] Hooks installed: %d/%d.", installedHooksCount, enabledHooksCount);
+    # FIX: Updated format specifiers to %zu to match size_t counters
+    LOG_INFO("[SDKRuntime] Hooks installed: %zu/%zu.", installedHooksCount, enabledHooksCount);
 
     HookManager::instance().enableHooks();
